- 修正 bxslider js error - 隱藏 side nav +- icons - 微調 footer 連結列表

+++ assets/js/cms.js
@@ -35,7 +35,7 @@ function startSlider () {
     slider = startSlider();
 
     $(window).resize(function(){
-        if (typeof slider !== 'undefined') {
+        if (typeof slider !== 'undefined' && typeof slider.destroySlider === "function") {
             slider.destroySlider();
             slider = startSlider();
         }
